pyproject.toml
[project]
name = "defhack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29.0",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "mgrs>=1.5.0",
    "openai>=2.1.0",
    "packaging>=25.0",
    "python-telegram-bot[job-queue]>=22.5",
    "pillow>=10.4.0",
    "staticmap>=0.5.5",
    "numpy>=1.26",
    "pydantic>=2.11.9",
    "pillow>=10.0",
    "yolov5>=7.0.14",
    "open-clip-torch>=3.2.0",
    "matplotlib>=3.10.6",
    "tqdm>=4.67.1",
    "opencv-python>=4.11.0.86",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
